Added limited support for ++, --, += and -=
The implementation is rather limited and is aimed squarly at supporting
the forthcoming for loop (i++, i += 2  type of thing). Specifically:

-It only works with integers
-It only works with simple values, so the following won't work:
  - user.PowerLevel++
  - user.Ranks[4] += 10

There are many other cases where it won't behave as expected. The idea of
writing to variables from a template doesn't strike me as particularly useful.
That is, except for simple cases, such as for loop and variable declarations.