Allow numbers in variable names
Assignment tag supports both := and = assignment operators. However, they both
behave the same: overwriting existing values or definition undefined values.
The difference is that a message will be logged if the operator is used
inappropriately (it follow's Go's rules, := can be used if 1 or many fields
are new, else = should be used...)