Template inheritance.
Content can be captured via the <% content NAME { %>...<% } %> tag and
exported via the built-in yield method. Yield is a real core.Value, and thus can
be used in an assignment or any other place where a core.Value is supported.

Support for the default yield (which dump any content not in an explicit
content tag) is pretty hackish and done at the parser level. This is
unfortunate.